Protecting resources/endpoint by leveraging Auth

+++ project.py
@@ -142,6 +142,16 @@ def createUserIfNotExist(login_session):
         return createUser(login_session)
     else:
         return user_id
+
+def is_authenticated():
+    return 'username' in login_session
+
+def is_guest():
+    return 'username' not in login_session
+
+def is_owner(an_object):
+    return an_object.user_id == login_session['user_id']
+
 # Logout : https://github.com/udacity/ud330/blob/master/Lesson2/step6/project.py
 
 @app.route('/gdisconnect')
@@ -198,8 +208,8 @@ def restaurantsJSON():
 @app.route('/')
 @app.route('/restaurant/')
 def showRestaurants():
-  restaurants = session.query(Restaurant).order_by(asc(Restaurant.name))
-  return render_template('restaurants.html', restaurants = restaurants)
+  restaurants = session.query(Restaurant).order_by(asc(Resta  urant.name))
+  return render_template('restaurants.html' if 'username' not in login_session else 'publicrestaurants.html', restaurants = restaurants)
 
 #Create a new restaurant
 @app.route('/restaurant/new/', methods=['GET','POST'])
@@ -234,6 +244,10 @@ def editRestaurant(restaurant_id):
 @app.route('/restaurant/<int:restaurant_id>/delete/', methods = ['GET','POST'])
 def deleteRestaurant(restaurant_id):
   restaurantToDelete = session.query(Restaurant).filter_by(id = restaurant_id).one()
+  if (is_guest()):
+      return redirect('/login')
+  if (!is_owner(restaurantToDelete)):
+      return "Hello.. You are not the owner of this restaurant"        
   if request.method == 'POST':
     session.delete(restaurantToDelete)
     flash('%s Successfully Deleted' % restaurantToDelete.name)
@@ -247,9 +261,12 @@ def deleteRestaurant(restaurant_id):
 @app.route('/restaurant/<int:restaurant_id>/menu/')
 def showMenu(restaurant_id):
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+    owner = getUserInfo(restaurant.user_id)
     items = session.query(MenuItem).filter_by(restaurant_id = restaurant_id).all()
-    return render_template('menu.html', items = items, restaurant = restaurant)
-
+    is_public = 'username' not in login_session or owner.id
+                           != login_session['user_id']
+    return render_template(('publicmenu.html' if is_public else 'menu.html'
+                           ), items=items, restaurant=restaurant, creator= None if is_public else owner)
 
 
 #Create a new menu item
